+++ lib/listener.js
@@ -14,6 +14,7 @@ const logger = require('./logger')
 const rabbitmq = require('./rabbitmq')
 const sinceMap = require('./since-map')
 const datadog = require('./datadog')
+const dockerUtils = require('./docker.utils')
 
 module.exports = class EventListener {
   /**
@@ -94,7 +95,7 @@ module.exports = class EventListener {
           resolve()
         })
 
-      this.docker.testEvent()
+      dockerUtils.testEvent(this.docker)
     })
     .timeout(process.env.EVENT_TIMEOUT_MS)
     .catch(Promise.TimeoutError, () => {
