+++ lib/swarm.js
@@ -2,10 +2,7 @@
 require('loadenv')()
 
 const Promise = require('bluebird')
-const isEmpty = require('101/is-empty')
 const SwarmClient = require('@runnable/loki').Swarm
-
-const datadog = require('./datadog')
 const log = require('./logger')()
 
 module.exports = class Swarm extends SwarmClient {
@@ -34,39 +31,5 @@ module.exports = class Swarm extends SwarmClient {
           return nodes[key]
         })
       })
-  }
-
-  /**
-   * Do docker top on a random container.
-   * This should emit the `top` event.
-   * We ignore errors since we will restart if event failed
-   * @param  {Function} cb
-   */
-  testEvent () {
-    log.info('Swarm.prototype.testEvent')
-    const listTimer = datadog.timer('listContainers.time')
-    return this.listContainersAsync({
-      filters: {
-        state: ['running']
-      }
-    })
-    .then((containers) => {
-      listTimer.stop()
-      if (isEmpty(containers)) {
-        throw new Error('no running containers found')
-      }
-      var testContainerId = containers[0].Id
-      log.trace({ container: testContainerId }, 'testEvent - got container')
-
-      const topTimer = datadog.timer('top.time')
-      return this.topContainerAsync(testContainerId)
-        .then(() => {
-          topTimer.stop()
-          log.trace({ container: testContainerId }, 'testEvent - top successful')
-        })
-    })
-    .catch((err) => {
-      log.error({ err: err }, 'testEvent - ' + err.message)
-    })
   }
 }
