+++ test/unit/docker.utils.js
@@ -8,9 +8,10 @@ const Lab = require('lab')
 const sinon = require('sinon')
 require('sinon-as-promised')(Promise)
 const BaseDockerClient = require('@runnable/loki')._BaseClient
-const DockerClient = require('@runnable/loki').Docker
+const SwarmClient = require('@runnable/loki').Swarm
 
-const Docker = require('../../lib/docker')
+const Swarm = require('../../lib/swarm')
+const dockerUtils = require('../../lib/docker.utils')
 
 const lab = exports.lab = Lab.script()
 
@@ -19,25 +20,25 @@ const beforeEach = lab.beforeEach
 const describe = lab.experiment
 const it = lab.test
 
-describe('docker unit test', () => {
+describe('swarm unit test', () => {
   let docker
   const testHost = '10.0.0.1:4242'
 
   beforeEach((done) => {
-    docker = new Docker(testHost)
+    docker = new Swarm(testHost)
     done()
   })
 
   describe('testEvent', () => {
     beforeEach((done) => {
-      sinon.stub(DockerClient.prototype, 'listContainersAsync')
+      sinon.stub(SwarmClient.prototype, 'listContainersAsync')
       sinon.stub(BaseDockerClient.prototype, 'topContainerAsync').resolves({})
       sinon.stub(errorCat, 'report')
       done()
     })
 
     afterEach((done) => {
-      DockerClient.prototype.listContainersAsync.restore()
+      SwarmClient.prototype.listContainersAsync.restore()
       BaseDockerClient.prototype.topContainerAsync.restore()
       errorCat.report.restore()
       done()
@@ -45,29 +46,29 @@ describe('docker unit test', () => {
 
     it('should ignore error on list fail', (done) => {
       const testErr = new Error('calamitous')
-      DockerClient.prototype.listContainersAsync.rejects(testErr)
-      docker.testEvent().asCallback(done)
+      SwarmClient.prototype.listContainersAsync.rejects(testErr)
+      dockerUtils.testEvent(docker).asCallback(done)
     })
 
     it('should ignore error on empty containers', (done) => {
-      DockerClient.prototype.listContainersAsync.resolves([])
-      docker.testEvent().asCallback(done)
+      SwarmClient.prototype.listContainersAsync.resolves([])
+      dockerUtils.testEvent(docker).asCallback(done)
     })
 
     it('should ignore error top fail', (done) => {
       const testErr = new Error('grievous')
-      DockerClient.prototype.listContainersAsync.resolves([{Id: 1}])
+      SwarmClient.prototype.listContainersAsync.resolves([{Id: 1}])
       BaseDockerClient.prototype.topContainerAsync.rejects(testErr)
-      docker.testEvent().asCallback(done)
+      dockerUtils.testEvent(docker).asCallback(done)
     })
 
     it('should call docker with correct opts', (done) => {
       const testId = 'heinous'
-      DockerClient.prototype.listContainersAsync.resolves([{Id: testId}])
-      docker.testEvent().asCallback((err) => {
+      SwarmClient.prototype.listContainersAsync.resolves([{Id: testId}])
+      dockerUtils.testEvent(docker).asCallback((err) => {
         if (err) { return done(err) }
-        sinon.assert.calledOnce(DockerClient.prototype.listContainersAsync)
-        sinon.assert.calledWith(DockerClient.prototype.listContainersAsync, {
+        sinon.assert.calledOnce(SwarmClient.prototype.listContainersAsync)
+        sinon.assert.calledWith(SwarmClient.prototype.listContainersAsync, {
           filters: {
             state: ['running']
           }
@@ -79,4 +80,4 @@ describe('docker unit test', () => {
       })
     })
   }) // end testEvent
-}) // end testEvent
+})
